split tests into a separate file
